refactor(polytomy): deduplicate find_polytomy_nodes, use remove_node_if_trivial
- Extract generic find_polytomy_nodes to topology_cleanup/polytomy_nodes.rs
- Replace remove_obsolete_nodes with remove_node_if_trivial from reroot.rs
- remove_node_if_trivial properly sums branch lengths into merged edge
