+++ packages/treetime/src/representation/algo/topology_cleanup/mod.rs
@@ -3,4 +3,5 @@ mod __tests__;
 
 pub mod collapse;
 pub mod merge_shared_mutations;
+pub mod polytomy_nodes;
 pub mod reroot;
